% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_edf.R
\name{read_edf}
\alias{read_edf}
\title{Read EEG signal from an EDF file}
\usage{
read_edf(edf_file)
}
\arguments{
\item{edf_file}{EDF file}
}
\value{
List containing the EDF file items
}
\description{
Read EEG signal from an EDF file
}
